+++ src/sdk/pynni/nni/nas/pytorch/search_space_zoo/__init__.py
@@ -1,2 +1,3 @@
 from .darts_cell import DartsCell
-
+from .enas_cell import ENASMicroLayer
+from .enas_cell import ENASReductionLayer
